+++ pom.xml
@@ -27,7 +27,7 @@
     <modules>
         <module>component</module>
     </modules>
-    <groupId>org.wso2.extension.siddhi.io.nats</groupId>
+    <groupId>io.siddhi.extension.io.nats</groupId>
     <artifactId>siddhi-io-nats-parent</artifactId>
     <version>2.0.3-SNAPSHOT</version>
     <name>Siddhi IO NATS Extension Parent</name>
@@ -82,13 +82,13 @@
                 </exclusions>
             </dependency>
             <dependency>
-                <groupId>org.wso2.extension.siddhi.map.xml</groupId>
+                <groupId>io.siddhi.extension.map.xml</groupId>
                 <artifactId>siddhi-map-xml</artifactId>
                 <version>${siddhi.map.xml.version}</version>
                 <scope>test</scope>
             </dependency>
             <dependency>
-                <groupId>org.wso2.extension.siddhi.map.json</groupId>
+                <groupId>io.siddhi.extension.map.json</groupId>
                 <artifactId>siddhi-map-json</artifactId>
                 <version>${siddhi.map.json.version}</version>
                 <scope>test</scope>
@@ -137,13 +137,13 @@
         </plugins>
     </build>
     <properties>
-        <siddhi.version>5.0.0-m11</siddhi.version>
+        <siddhi.version>5.0.0</siddhi.version>
         <siddhi.version.range>[5.0.0, 6.0.0)</siddhi.version.range>
         <siddhi.doc.gen.version>4.5.0</siddhi.doc.gen.version>
         <log4j.version>1.2.17.wso2v1</log4j.version>
         <carbon.transport.version>4.4.15</carbon.transport.version>
-        <siddhi.map.json.version>5.0.1</siddhi.map.json.version>
-        <siddhi.map.xml.version>5.0.0</siddhi.map.xml.version>
+        <siddhi.map.json.version>5.0.2</siddhi.map.json.version>
+        <siddhi.map.xml.version>5.0.2</siddhi.map.xml.version>
         <jacoco.plugin.version>0.7.9</jacoco.plugin.version>
         <nats.version>2.1.0</nats.version>
         <surefire.plugin.version>2.22.0</surefire.plugin.version>
